+++ 03-frontend/angular-ecommerce/src/app/services/cart.service.ts
@@ -6,7 +6,7 @@ import { Subject } from 'rxjs';
   providedIn: 'root'
 })
 export class CartService {
-
+  
   cartItems: CartItem[] = [];
 
   // Subject is a subclass of Observable
@@ -86,4 +86,31 @@ export class CartService {
       console.log(`totalPrice: ${totalPriceValue.toFixed(2)}, totalQuantity: ${totalQuantityValue}`);
       console.log('-------------')
   }
+
+  // decrement an item in the cart
+  decrementQuantity(theCartItem: CartItem) {
+    theCartItem.quantity--;
+
+    if (theCartItem.quantity === 0) {
+      this.remove(theCartItem)
+    }
+    else {
+      this.computeCartTotals();
+    }
+  }
+
+  // remove an item from the cart
+  remove(theCartItem: CartItem) {
+    
+    // get index of item in the array
+    const itemIndex = this.cartItems.findIndex( tempCartItem => tempCartItem.id === theCartItem.id );
+
+    // if found, remove the item from the array at the given index
+    if (itemIndex > -1) {
+      this.cartItems.splice(itemIndex, 1);
+
+      this.computeCartTotals();
+    }
+  }
+
 } // end CartService